Update enet_packet_destroy documentation to reflect when it should be used.
I was working on a project and kept getting a segfault until i read deeper into the docs to know that `enet_packet_send` takes ownership of the packet. It might be beneficial to document that nearer the destroy function, but honestly I'm not sure the best way to clarify.